pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dvc>=3.66.1",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",
    "catboost>=1.2.7",
    "mlflow>=2.17.0",
    "matplotlib>=3.9.0",
    "joblib>=1.4.0",
    "pyyaml>=6.0.2",
    "pyarrow>=17.0.0",
]
